refactor: Controller 형식 통일, 인증 기능 추가
