+++ backend/src/main/java/com/festago/bookmark/presentation/v1/ArtistBookmarkV1Controller.java
@@ -1,6 +1,7 @@
-package com.festago.bookmark.presentation;
+package com.festago.bookmark.presentation.v1;
 
 import com.festago.auth.annotation.Member;
+import com.festago.auth.annotation.MemberAuth;
 import com.festago.bookmark.application.ArtistBookmarkV1QueryService;
 import com.festago.bookmark.dto.v1.ArtistBookmarkV1Response;
 import io.swagger.v3.oas.annotations.Operation;
@@ -14,15 +15,18 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@Tag(name = "아티스트 북마크 V1")
 @RequestMapping("/api/v1/bookmarks/artists")
+@Tag(name = "아티스트 북마크 V1")
 public class ArtistBookmarkV1Controller {
 
     private final ArtistBookmarkV1QueryService artistBookmarkV1QueryService;
 
+    @MemberAuth
     @GetMapping
     @Operation(description = "유저의 아티스트 북마크 목록을 조회한다.", summary = "아티스트 북마크 조회")
-    public ResponseEntity<List<ArtistBookmarkV1Response>> findArtistBookmarksByMemberId(@Member Long memberId) {
+    public ResponseEntity<List<ArtistBookmarkV1Response>> findArtistBookmarksByMemberId(
+        @Member Long memberId
+    ) {
         return ResponseEntity.ok(artistBookmarkV1QueryService.findArtistBookmarksByMemberId(memberId));
     }
 }
