+++ backend/src/main/java/com/festago/bookmark/presentation/v1/SchoolBookmarkV1Controller.java
@@ -1,6 +1,7 @@
-package com.festago.bookmark.presentation;
+package com.festago.bookmark.presentation.v1;
 
 import com.festago.auth.annotation.Member;
+import com.festago.auth.annotation.MemberAuth;
 import com.festago.bookmark.application.SchoolBookmarkV1QueryService;
 import com.festago.bookmark.dto.v1.SchoolBookmarkV1Response;
 import io.swagger.v3.oas.annotations.Operation;
@@ -15,11 +16,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/bookmarks/schools")
-@Tag(name = "학교 북마크 API V1")
+@Tag(name = "학교 북마크 V1")
 public class SchoolBookmarkV1Controller {
 
     private final SchoolBookmarkV1QueryService schoolBookmarkV1QueryService;
 
+    @MemberAuth
     @GetMapping
     @Operation(description = "특정한 회원의 학교 북마크 목록을 반환한다", summary = "회원 학교 북마크 목록 조회")
     public ResponseEntity<List<SchoolBookmarkV1Response>> findAllByMemberId(@Member Long memberId) {
